Refactor `Rebalancing` service interface for testing purposes
